Remove meta_path hack and poetry_install code. Im sick of the bugs and hackiness for it. Guides need updated to instruct people how to install dependencies!
