+++ sickchill/update_manager/pip.py
@@ -1,9 +1,11 @@
-from typing import Union
+import subprocess
+import sys
+from pathlib import Path
+from typing import List, Union
 
 from packaging import version as packaging_version
 
-from sickchill import settings, version
-from sickchill.init_helpers import pip_install
+from sickchill import logger, settings, version
 from sickchill.oldbeard import helpers, notifiers
 
 from .abstract import UpdateManagerBase
@@ -70,9 +72,81 @@ class PipUpdateManager(UpdateManagerBase):
         return self.get_newest_version() > self.get_current_version()
 
     def update(self):
-        if not pip_install("sickchill"):
+        if not self.pip_install("sickchill"):
             return False
 
         settings.CUR_COMMIT_HASH = self._newest_version
         notifiers.notify_git_update(f"{settings.CUR_COMMIT_HASH or ''}")
+        return True
+
+    def pip_install(self, packages: Union[List[str], str]) -> bool:
+
+        sickchill_module = Path(__file__).parent.resolve()
+        pyproject_path = sickchill_module.parent / "pyproject.toml"
+
+        def subprocess_call(cmd_list):
+            try:
+                process = subprocess.Popen(
+                    cmd_list, stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True, cwd=os.getcwd()
+                )
+                stdout, stderr = process.communicate()
+                process.wait()
+                if stdout or stderr:
+                    logger.info(f"Command result: {stdout or stderr}")
+            except Exception as error:
+                logger.info(f"Unable to run command: {error}")
+                return 126
+            return process.returncode
+
+        def get_os_id():
+            os_release = Path("/etc/os-release").resolve()
+            if os_release.is_file():
+                from configparser import ConfigParser
+
+                parser = ConfigParser()
+                parser.read_string("[DEFAULT]\n" + os_release.read_text())
+                try:
+                    return parser["DEFAULT"]["ID"]
+                except (KeyError, IndexError):
+                    pass
+
+        cmd = [
+            sys.executable,
+            "-m",
+            "pip",
+            "install",
+            "--no-input",
+            "--disable-pip-version-check",
+            "--no-python-version-warning",
+            "--no-color",
+            # "--trusted-host=pypi.org",
+            # "--trusted-host=files.pythonhosted.org",
+            "-qU",
+        ]
+
+        os_id = get_os_id()
+        if os_id in ("alpine", "ubuntu"):
+            cmd.append(f"--find-links=https://wheel-index.linuxserver.io/{os_id}/")
+
+        if os_id == "alpine":
+            cmd.append(f"--extra-index-url=https://alpine-wheels.github.io/index")
+
+        elif os_id in ("raspian", "osmc"):
+            cmd.append(f"--extra-index-url=https://www.piwheels.org/simple")
+
+        syno_wheelhouse = pyproject_path.parent.with_name("wheelhouse")
+        if syno_wheelhouse.is_dir():
+            logger.debug(f"Found wheelhouse dir at {syno_wheelhouse}")
+            cmd.append(f"-f{syno_wheelhouse}")
+
+        cmd += packages
+
+        logger.debug(f"pip args: {' '.join(cmd)}")
+
+        result = subprocess_call(cmd)
+        if result != 0:  # Not Ok
+            logger.info("Trying user site-packages")
+            result = subprocess_call(cmd + ["--user"])
+            if result != 0:  # Not Ok
+                return False
         return True
